Added the mixed-speed test-scripts
